gs201: install libOpenCL
This required disabling the GPU-related PRODUCT_SOONG_NAMESPACES so that
we don't get an installed-path collision between the prebuilt gs201
libOpenCL and the gs101 libOpenCL that we build from source. Once we're
able to build both gs101 and gs201 from the same source (b/185349920)
we'll need to re-add the namespaces.

Bug: 190424017
Test: gfx-score
Change-Id: I9095e099a120494fab0c33fc6bd1ace2c69f3bc4