Use modprobe --all to load all modules in modules.load
Previously, we ran modprobe command for every .ko file listed in
init.insmod.*.cfg. Let's move to a different approach where we call
modprobe with the following command line.

modprobe -b --all=/vendor/lib/modules/modules.load

This will insmod all modules listed in modules.load except for those
that are mentioned in the blocklist at
/vendor/lib/modules/modules.blocklist

A common reason for a module to be on the blocklist is that it must only
be loaded under a certain condition like when a specific service is
launched.

Bug: 192241728
Change-Id: I518426308d0bfc358c8c815a69d431e27ebcf6e4