gs201: add compatibility path for TPU
This is to expose a common path to runtime. This is the gs201 version of the WCH change ag/15100070. It alias /dev/janeiro to /dev/edgetpu. This is a temporary measure to make /dev/edgetpu work. Once the TPU kernel driver is refactored to produce /dev/edgetpu by default, we'll remove this symlink.

Test: Manual. Verified edgetpu_logging works with modified path
Bug: 191193044

Change-Id: Icc70264c0c7d20ea7b6aa43903e109d9180dd022