Build init_boot for new devices
Builds a new init_boot.img with the ramdisk that was previously in
boot.img.
Removes the ramdisk from boot.img.
Change fstab to make sure the /dev/block/by-name/init_boot_* entries are
created in first stage init for Verified Boot.
Add init_boot.img to the list of OTA images so it is picked up for OTA
updates.

Test: unpack_bootimg --boot_img=boot.img
Test: unpack_bootimg --boot_img=init_boot.img
Test: Build and boot new device with required bootloader changes
Bug: 211823573

Change-Id: Idbd700b17d862f48ea568d1274a6550b965763ea