Upgrade android.hardware.dumpstate from HIDL 1.1 to AIDL 1
lshal is not supported for AIDL, now use dumpsys instead.
Update the debug command as below.
  Old: lshal debug android.hardware.dumpstate@1.1::IDumpstateDevice/default [section]
  New: dumpsys android.hardware.dumpstate.IDumpstateDevice/default [section]

Currently dumpsys does not start the Lazy HAL service, only dumpstate does.
Because we need to run dumpsys for debugging, keep the dumpstate HAL running at boot.
Do not set it to be a lazy HAL.

Test: atest VtsHalDumpstateTargetTest pass
      adb shell dumpsys -t 30 android.hardware.dumpstate.IDumpstateDevice/default all
Bug: 223118410
Change-Id: I7f866a57c3eff8c9783fee89dce205cf9728c459