pkvm: Prefix metrics when running on A0 chip
A0 and B0 chips have different performance characteristics when it comes
to the hypervisor because B0 waits for SysMMU_SYNCs to complete during
every memory ownership transaction.

Set debug.hypervisor.metrics_tag=hwrev_a0 on PROTO1.x and EVT1.0
devices. EVT1.1 and later are B0 and benchmarks are representative of
production devices. This is only done on userdebug/eng builds.

Test: 'getprop | grep hypervisor' on A0/B0 devices
Change-Id: I0c81a0e4706a13cd63358de23048087a9190b9da