Set activelyPreferBadwifi config to 1 globally for gs201 devices
This flag will not take effect if the system is configured to
avoid bad wifis. Set it to 1 globally. Then it will sprint
into effect for all mcc/mncs where config_networkAvoidBadWifi is 0.

For carriers that don't need this behavior change but
config_networkAvoidBadWifi is 0, set activelyPreferBadwifi
to 0 for their mccmncs.

Test: manual
Bug: 267489867
Change-Id: I4b6fe4964e8fd5d61978f1433dcbbd9e9b10ffc3