gs201: support splitting modules between system/vendor_dlkm
Currently, all gs201 targets have a system_dlkm image but their
modules are stored on the vendor_dlkm partition. In order to
support the staging kernel, modules have to be split across the
system/vendor_dlkm images.

This fix enables targets to define modules on the system_dlkm image.
If so, the system/vendor_dlkm split will be honored. Otherwise, all
modules will remain stored on the vendor_dlkm image.

Test:
- Build and flash aosp_panther, no errors.
- Build and flash a target defining BOARD_SYSTEM_KERNEL_MODULES, ensure
  modules are correctly split across the two dlkm partitions.

Bug: 285177263
Change-Id: I80131d97073ce666a9228a47c2fff94b76139a64
Signed-off-by: Samuel Gosselin <sgosselin@google.com>