suspend: enable adaptive suspend
This CL adds build properties enabling Adaptive Suspend's short
suspend backoff with a short suspend threshold of 2 seconds and
a max backoff time of 40s.  This is the same configuration that
was deployed on the previous generation of devices.

The plan is to monitor suspend behavior, screen-off uptime,
wakeup rates, and the distribution of screen-off mA on
these devices in comparison to those same attributes on late
UDC builds, which will inform whether we ship or revert.

Bug: 208947015
Test: booted with a test kernel and used alarmtimer mods to
exercise high-frequency wakeup behavior

Change-Id: Icf75c4dd705050833df5e48aaf0c28afa5754adb